Move the signup page to TS and functional components with hooks.
