+++ client/app.tsx
@@ -6,7 +6,7 @@ import { EmailVerificationUi } from './auth/email-verification'
 import { ForgotPassword, ForgotUser, ResetPassword } from './auth/forgot'
 import { LoggedInFilter } from './auth/logged-in-filter'
 import Login from './auth/login'
-import Signup from './auth/signup'
+import { Signup } from './auth/signup'
 import { ConnectedDialogOverlay } from './dialogs/connected-dialog-overlay'
 import { usePixelShover } from './dom/pixel-shover'
 import { DownloadPage } from './download/download-page'
